feat: 增加测试 #3594
